DEP/AWS/requirements.txt
boto3==1.28.53
paramiko==3.3.1
python-dotenv==1.0.0
